Simplify seekbar calculation and optimize change listener output

+++ app/src/main/java/com/cyb3rko/flashdim/seekbar/SeekBarChangeListener.kt
@@ -1,5 +1,5 @@
 package com.cyb3rko.flashdim.seekbar
 
 interface SeekBarChangeListener {
-    fun onProgressChanged(progress: Float)
+    fun onProgressChanged(progress: Int)
 }
